# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digilist Backoffice API
  description: |
    Complete API for the Digilist facility booking and management platform.

    ## Authentication

    Most endpoints require authentication via JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <jwt-token>
    ```

    Additional headers for tenant context:
    - `X-Tenant-Id`: UUID of the tenant (required for most endpoints)
    - `X-License-Key`: License key for API access

    ## Rate Limiting

    API requests are limited to 1000 requests per minute per tenant.

    ## Response Format

    All responses follow a standard format:

    **Success (single resource):**
    ```json
    { "data": { ... } }
    ```

    **Success (list):**
    ```json
    { "data": [...], "meta": { "total": 100, "page": 1, "limit": 20, "totalPages": 5 } }
    ```

    **Error:**
    ```json
    { "error": { "code": "ERROR_CODE", "message": "Human readable message" } }
    ```
  version: 1.0.0
  contact:
    name: Xala Technologies
    email: api@xalatechnologies.com
    url: https://xalatechnologies.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.digilist.no
    description: Production server
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      operationId: getBookings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/BookingStatus"
        - name: listingId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBookings"

    post:
      tags: [Bookings]
      summary: Create booking
      operationId: createBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingDTO"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Conflict (time slot not available)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking by ID
      operationId: getBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Bookings]
      summary: Update booking
      operationId: updateBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingDTO"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"

    delete:
      tags: [Bookings]
      summary: Delete booking
      operationId: deleteBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/bookings/{id}/confirm:
    put:
      tags: [Bookings]
      summary: Confirm pending booking
      operationId: confirmBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: Confirmed booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/bookings/{id}/cancel:
    put:
      tags: [Bookings]
      summary: Cancel booking
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/bookings/pricing:
    get:
      tags: [Bookings]
      summary: Calculate booking pricing
      operationId: calculatePricing
      security:
        - bearerAuth: []
      parameters:
        - name: listingId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: startTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Pricing calculation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BookingPricing"

  # ==========================================================================
  # Settings Endpoints
  # ==========================================================================
...
  /api/discount-codes:
    get:
      tags: [Discount Codes]
      summary: List discount codes
      operationId: getDiscountCodes
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: List of discount codes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DiscountCode"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

    post:
      tags: [Discount Codes]
      summary: Create discount code
      operationId: createDiscountCode
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscountCodeDTO"
      responses:
        "201":
          description: Created discount code
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DiscountCode"
        "409":
          description: Code already exists
  /api/discount-codes/{id}:
    get:
      tags: [Discount Codes]
      summary: Get discount code
      operationId: getDiscountCode
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Discount code
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DiscountCode"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Discount Codes]
      summary: Update discount code
      operationId: updateDiscountCode
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                value:
                  type: number
                isActive:
                  type: boolean
                maxUses:
                  type: integer
      responses:
        "200":
          description: Updated discount code

    delete:
      tags: [Discount Codes]
      summary: Delete discount code
      operationId: deleteDiscountCode
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/discount-codes/validate:
    post:
      tags: [Discount Codes]
      summary: Validate discount code
      operationId: validateDiscountCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: SUMMER25
                bookingValue:
                  type: number
                  example: 1000
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      valid:
                        type: boolean
                      discountAmount:
                        type: number
                      reason:
                        type: string

  # ==========================================================================
  # Health Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    BookingStatus:
      type: string
      enum: [pending, confirmed, cancelled, completed]
    PaymentStatus:
      type: string
      enum: [unpaid, paid, partial, refunded]
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # Auth Types
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/BookingStatus"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        totalPrice:
          type: string
        currency:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    BookingResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Booking"
    PaginatedBookings:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateBookingDTO:
      type: object
      required: [listingId, startTime, endTime]
      properties:
        listingId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type: string
    UpdateBookingDTO:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type: string
    BookingPricing:
      type: object
      properties:
        listingId:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        durationHours:
          type: number
        basePrice:
          type: number
        discount:
          type: number
        totalPrice:
          type: number
        currency:
          type: string

    # Other Types
...
    DiscountCode:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        isActive:
          type: boolean
        usedCount:
          type: integer
        maxUses:
          type: integer
        validFrom:
          type: string
          format: date
        validUntil:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    CreateDiscountCodeDTO:
      type: object
      required: [code, type, value]
      properties:
        code:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        maxUses:
          type: integer
        validFrom:
          type: string
          format: date
        validUntil:
          type: string
          format: date
security:
  - bearerAuth: []